# swagger.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/v1/branches":  {
                                           "post":  {
                                                        "operationId":  "BranchesController_create_v1",
                                                        "parameters":  [

                                                                       ],
                                                        "requestBody":  {
                                                                            "required":  true,
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/CreateBranchDto"
                                                                                                                                }
                                                                                                                 }
                                                                                        }
                                                                        },
                                                        "responses":  {
                                                                          "201":  {
                                                                                      "description":  "The branch has been successfully created."
                                                                                  }
                                                                      },
                                                        "security":  [
                                                                         {
                                                                             "access-token":  [

                                                                                              ]
                                                                         }
                                                                     ],
                                                        "summary":  "Create a new branch",
                                                        "tags":  [
                                                                     "Branches"
                                                                 ]
                                                    },
                                           "get":  {
                                                       "operationId":  "BranchesController_list_v1",
                                                       "parameters":  [
                                                                          {
                                                                              "name":  "page",
                                                                              "required":  false,
                                                                              "in":  "query",
                                                                              "description":  "Page number",
                                                                              "schema":  {
                                                                                             "minimum":  1,
                                                                                             "default":  1,
                                                                                             "type":  "number"
                                                                                         }
                                                                          },
                                                                          {
                                                                              "name":  "limit",
                                                                              "required":  false,
                                                                              "in":  "query",
                                                                              "description":  "Number of records per page",
                                                                              "schema":  {
                                                                                             "minimum":  1,
                                                                                             "maximum":  100,
                                                                                             "default":  20,
                                                                                             "type":  "number"
                                                                                         }
                                                                          },
                                                                          {
                                                                              "name":  "search",
                                                                              "required":  false,
                                                                              "in":  "query",
                                                                              "description":  "Search by branch name",
                                                                              "schema":  {
                                                                                             "example":  "Head",
                                                                                             "type":  "string"
                                                                                         }
                                                                          },
                                                                          {
                                                                              "name":  "isActive",
                                                                              "required":  false,
                                                                              "in":  "query",
                                                                              "description":  "Filter by active status",
                                                                              "schema":  {
                                                                                             "example":  true,
                                                                                             "type":  "boolean"
                                                                                         }
                                                                          }
                                                                      ],
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "List of branches with pagination."
                                                                                 }
                                                                     },
                                                       "security":  [
                                                                        {
                                                                            "access-token":  [

                                                                                             ]
                                                                        }
                                                                    ],
                                                       "summary":  "List all branches",
                                                       "tags":  [
                                                                    "Branches"
                                                                ]
                                                   }
                                       },
                  "/api/v1/branches/{id}":  {
                                                "get":  {
                                                            "operationId":  "BranchesController_getById_v1",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "required":  true,
                                                                                   "in":  "path",
                                                                                   "description":  "Branch ID",
                                                                                   "schema":  {
                                                                                                  "type":  "string"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "The branch details."
                                                                                      }
                                                                          },
                                                            "security":  [
                                                                             {
                                                                                 "access-token":  [

                                                                                                  ]
                                                                             }
                                                                         ],
                                                            "summary":  "Get branch by ID",
                                                            "tags":  [
                                                                         "Branches"
                                                                     ]
                                                        },
                                                "patch":  {
                                                              "operationId":  "BranchesController_update_v1",
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "id",
                                                                                     "required":  true,
                                                                                     "in":  "path",
                                                                                     "description":  "Branch ID",
                                                                                     "schema":  {
                                                                                                    "type":  "string"
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "requestBody":  {
                                                                                  "required":  true,
                                                                                  "content":  {
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/UpdateBranchDto"
                                                                                                                                      }
                                                                                                                       }
                                                                                              }
                                                                              },
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "The branch has been successfully updated."
                                                                                        }
                                                                            },
                                                              "security":  [
                                                                               {
                                                                                   "access-token":  [

                                                                                                    ]
                                                                               }
                                                                           ],
                                                              "summary":  "Update branch",
                                                              "tags":  [
                                                                           "Branches"
                                                                       ]
                                                          },
                                                "delete":  {
                                                               "operationId":  "BranchesController_delete_v1",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "id",
                                                                                      "required":  true,
                                                                                      "in":  "path",
                                                                                      "description":  "Branch ID",
                                                                                      "schema":  {
                                                                                                     "type":  "string"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "The branch has been successfully deleted."
                                                                                         }
                                                                             },
                                                               "security":  [
                                                                                {
                                                                                    "access-token":  [

                                                                                                     ]
                                                                                }
                                                                            ],
                                                               "summary":  "Soft delete a branch",
                                                               "tags":  [
                                                                            "Branches"
                                                                        ]
                                                           }
                                            },
                  "/api/v1/branches/{id}/status":  {
                                                       "patch":  {
                                                                     "operationId":  "BranchesController_updateStatus_v1",
                                                                     "parameters":  [
                                                                                        {
                                                                                            "name":  "id",
                                                                                            "required":  true,
                                                                                            "in":  "path",
                                                                                            "description":  "Branch ID",
                                                                                            "schema":  {
                                                                                                           "type":  "string"
                                                                                                       }
                                                                                        }
                                                                                    ],
                                                                     "requestBody":  {
                                                                                         "required":  true,
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/UpdateBranchStatusDto"
                                                                                                                                             }
                                                                                                                              }
                                                                                                     }
                                                                                     },
                                                                     "responses":  {
                                                                                       "200":  {
                                                                                                   "description":  "The branch status has been successfully updated."
                                                                                               }
                                                                                   },
                                                                     "security":  [
                                                                                      {
                                                                                          "access-token":  [

                                                                                                           ]
                                                                                      }
                                                                                  ],
                                                                     "summary":  "Update branch status (activate/deactivate)",
                                                                     "tags":  [
                                                                                  "Branches"
                                                                              ]
                                                                 }
                                                   }
              },
    "info":  {
                 "title":  "CRM + HRMS API",
                 "description":  "Production API documentation for the CRM + HRMS platform.",
                 "version":  "1.0.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "access-token":  {
                                                                    "scheme":  "bearer",
                                                                    "bearerFormat":  "JWT",
                                                                    "type":  "http",
                                                                    "description":  "Enter JWT access token"
                                                                }
                                           },
                       "schemas":  {
...
                                       "CreateBranchDto":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "name":  {
                                                                                               "type":  "string",
                                                                                               "description":  "Name of the branch or location",
                                                                                               "example":  "Headquarters"
                                                                                           },
                                                                                  "address":  {
                                                                                                  "type":  "string",
                                                                                                  "description":  "Physical address of the branch",
                                                                                                  "example":  "123 Main St, New York, NY 10001"
                                                                                              }
                                                                              },
                                                               "required":  [
                                                                                "name"
                                                                            ]
                                                           },
                                       "UpdateBranchDto":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "name":  {
                                                                                               "type":  "string",
                                                                                               "description":  "Name of the branch or location",
                                                                                               "example":  "Headquarters"
                                                                                           },
                                                                                  "address":  {
                                                                                                  "type":  "string",
                                                                                                  "description":  "Physical address of the branch",
                                                                                                  "example":  "123 Main St, New York, NY 10001"
                                                                                              }
                                                                              }
                                                           },
                                       "UpdateBranchStatusDto":  {
                                                                     "type":  "object",
                                                                     "properties":  {
                                                                                        "isActive":  {
                                                                                                         "type":  "boolean",
                                                                                                         "description":  "Active status of the branch",
                                                                                                         "example":  true
                                                                                                     }
                                                                                    },
                                                                     "required":  [
                                                                                      "isActive"
                                                                                  ]
                                                                 }
                                   }
                   }
}